.github/ACDbot/requirements.txt
click==8.1.7
google_api_python_client==2.157.0
google_auth_oauthlib==1.2.1
protobuf>=3.18.3
PyGithub==2.5.0
pytest==8.3.4
pytz==2022.1
Requests==2.32.3
python-dateutil>=2.8.2
python-telegram-bot==20.3
webvtt-py==0.4.6



